Revert "fixing a bunch of warnings in visual studio 2010"
This reverts commit 351ac31861b0d32f51b8f098ee38e25be1c52bfc.

+++ Source/MediaStorageAndFileFormat/gdcmJPEG2000Codec.cxx
@@ -376,7 +376,7 @@ bool JPEG2000Codec::Decode(std::istream &is, std::ostream &os)
   // FIXME: Do some stupid work:
   is.seekg( 0, std::ios::end);
   std::streampos buf_size = is.tellg();
-  char *dummy_buffer = new char[(unsigned int)buf_size];
+  char *dummy_buffer = new char[buf_size];
   is.seekg(0, std::ios::beg);
   is.read( dummy_buffer, buf_size);
   unsigned char *src = (unsigned char*)dummy_buffer;
@@ -1083,10 +1083,10 @@ bool JPEG2000Codec::GetHeaderInfo(std::istream &is, TransferSyntax &ts)
   // FIXME: Do some stupid work:
   is.seekg( 0, std::ios::end);
   std::streampos buf_size = is.tellg();
-  char *dummy_buffer = new char[(unsigned int)buf_size];
+  char *dummy_buffer = new char[buf_size];
   is.seekg(0, std::ios::beg);
   is.read( dummy_buffer, buf_size);
-  bool b = GetHeaderInfo( dummy_buffer, (size_t)buf_size, ts );
+  bool b = GetHeaderInfo( dummy_buffer, buf_size, ts );
   delete[] dummy_buffer;
   return b;
 }
